backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
cloudinary==1.37.0
tensorflow-cpu==2.13.0
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
opencv-python==4.8.1.78
motor==3.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.5.0
email-validator==2.3.0
